@@ -2,7 +2,6 @@
     "name": "kohanaworld/hmmmath",
     "description": "Collection of math related PHP functions (PHP >=7.2 <=8.3",
     "license": "MIT",
-    "version": "0.9.0",
     "authors": [
         {
             "name": "Lars Strojny",
@@ -24,9 +23,11 @@
         }
     },
     "require": {
-        "php": "^7.3|^8.0"
+        "php": "^7.4|^8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^9"
+        "phpunit/phpunit": "^9",
+        "phpstan/phpstan": "^1.11",
+        "squizlabs/php_codesniffer": "3.*"
     }
 }
