@@ -5,50 +5,70 @@ use ArrayAccess;
 use Iterator;
 use hmmmath\Exception\BadMethodCallException;
 
+
+/**
+ * @implements Iterator<int, FibonacciNumber>
+ * @implements ArrayAccess<int, FibonacciNumber>
+ */
 class FibonacciSequence implements Iterator, ArrayAccess
 {
     /** @var FibonacciNumber */
-    private $number;
+    private FibonacciNumber $number;
 
     /** @var FibonacciNumber */
-    private $initial;
+    private FibonacciNumber $initial;
 
     /** @var int */
-    private $key = 0;
+    private int $key = 0;
 
     public function __construct(int $start = 0, int $increment = 1)
     {
         $this->number = $this->initial = new FibonacciNumber($start, $increment);
     }
 
-    public function current(): int
+	/**
+	 * @return int
+	 */
+	public function current(): int
     {
         return $this->number->getCurrent();
     }
 
-    public function key(): int
+	/**
+	 * @return int
+	 */
+	public function key(): int
     {
         return $this->key;
     }
 
-    public function valid(): bool
+	/**
+	 * @return bool
+	 */
+	public function valid(): bool
     {
         return true;
     }
 
-    public function rewind(): void
+	/**
+	 * @return void
+	 */
+	public function rewind(): void
     {
         $this->key = 0;
         $this->number = $this->initial;
     }
 
-    public function next(): void
+	/**
+	 * @return void
+	 */
+	public function next(): void
     {
         ++$this->key;
         $this->number = $this->number->getNext();
     }
 
-    public function offsetExists($offset): bool // @codingStandardsIgnoreLine
+	public function offsetExists($offset): bool // @codingStandardsIgnoreLine
     {
         return true;
     }
@@ -64,7 +84,7 @@ class FibonacciSequence implements Iterator, ArrayAccess
         return $number->getCurrent();
     }
 
-    public function offsetSet($offset, $value): void // @codingStandardsIgnoreLine
+	public function offsetSet($offset, $value): void // @codingStandardsIgnoreLine
     {
         throw BadMethodCallException::invalidMethod(__CLASS__, __FUNCTION__);
     }
