@@ -3,7 +3,7 @@ namespace hmmmath\Exception;
 
 use InvalidArgumentException as BaseInvalidArgumentException;
 
-class InvalidArgumentException extends BaseInvalidArgumentException
+final class InvalidArgumentException extends BaseInvalidArgumentException
 {
     /**
      * @param mixed $actualValue
